requirements.txt
aiohttp
aiosignal
annotated-types
anyio
attrs
blinker
certifi
charset-normalizer
click
dataclasses-json
diskcache
Flask
Flask-Cors
frozenlist
h11
httpcore
httpx
idna
itsdangerous
Jinja2
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-ollama
langchain-text-splitters
langsmith
MarkupSafe
marshmallow
multidict
mypy-extensions
numpy
ollama
orjson
packaging
propcache
pydantic
pydantic-settings
pydantic_core
pydub
PyPDF2
python-dotenv
PyYAML
requests
requests-toolbelt
sniffio
SQLAlchemy
tenacity
typing-inspect
typing_extensions
urllib3
Werkzeug
yarl
ffmpeg-python
SpeechRecognition
openai
